Extracts the PluginsNavigationHeader from PluginsBrowser
